userEndpoints: favourite, cart

+++ DTO/Product/ProductDtoBaseInfo.cs
@@ -15,7 +15,7 @@ namespace MagazinchikAPI.DTO
 
         public long ReviewCount { get; set; }
 
-        public long AverageRating { get; set; }
+        public float AverageRating { get; set; }
 
         public long Purchases { get; set; }
     }
